refactor(dns): extract the resolve core into QueryPipeline + ReplyCapture (#911)
Pure move, zero behavior change — the prerequisite for non-UDP DNS
transports (#910).

- handle_query and its helpers move from server.rs into a QueryPipeline
  struct in pipeline.rs; the UDP listener drives the same bodies via
  pipeline.handle(packet, src, reply, client). The DnsSocket send seam
  stays — the raw-byte relay paths in forward_via_tunnel /
  forward_via_conditional have no parsed Message on their parse-failure
  branch, so a -> Response shape would put the REFUSED/negative/
  rebinding semantics at risk.
- New ReplyCapture: a capacity-1 mpsc DnsSocket impl so stream
  transports get "query bytes in → response bytes out" (every pipeline
  path sends exactly one response).
- New ClientIdentity: Ip(IpAddr) — today's snapshot-attributed behavior
  for UDP/LAN — vs Device { device_id, token } for SNI-attributed
  transports (the DoT listener, next sub-issue).
- UdpDnsServer now holds an Arc<QueryPipeline>; the UDP path is
  byte-identical (same function bodies, same order of operations), and
  server.rs keeps the lifecycle, the resolver/recursor builders, and
  the latency prober.
- dns/tests/server.rs passes unchanged (the historical
  crate::dns::server:: paths are re-exported); new tests cover
  ReplyCapture multi-send tolerance and a dropped receiver.
