+++ source/daemon/crates/wardnetd/src/dns/mod.rs
@@ -1,3 +1,5 @@
+pub mod pipeline;
+pub mod reply_capture;
 pub mod server;
 
 mod rate_limit;
