+++ source/daemon/crates/wardnetd/src/dns/tests/mod.rs
@@ -5,4 +5,5 @@
 // exercise the lifecycle and the `record_query` / `duration_to_ms` /
 // `upstream_label` helpers that don't need a fake `DnsSocket` plumbed.
 
+mod reply_capture;
 mod server;
